+++ packages/common/src/pipes/i18n_plural_pipe.ts
@@ -10,7 +10,7 @@ import {Pipe, PipeTransform} from '@angular/core';
 
 import {getPluralCategory, NgLocalization} from '../i18n/localization';
 
-import {invalidPipeArgumentError} from './invalid_pipe_argument_error';
+import {invalidPipeArgumentError} from './utils';
 
 const _INTERPOLATION_REGEXP: RegExp = /#/g;
