+++ packages/common/src/pipes/json_pipe.ts
@@ -6,7 +6,8 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {isSignal, Pipe, PipeTransform} from '@angular/core';
+import {Pipe, PipeTransform} from '@angular/core';
+import {warnIfSignal} from './utils';
 
 /**
  * @ngModule CommonModule
@@ -34,9 +35,7 @@ export class JsonPipe implements PipeTransform {
    * @param value A value of any type to convert into a JSON-format string.
    */
   transform(value: any): string {
-    if (ngDevMode && isSignal(value)) {
-      console.warn(`The JsonPipe does not unwrap signals. Received a signal with value:`, value());
-    }
+    ngDevMode && warnIfSignal('JsonPipe', value);
 
     return JSON.stringify(value, null, 2);
   }
