+++ packages/common/src/pipes/number_pipe.ts
@@ -18,7 +18,7 @@ import {
 import {formatCurrency, formatNumber, formatPercent} from '../i18n/format_number';
 import {getCurrencySymbol} from '../i18n/locale_data_api';
 
-import {invalidPipeArgumentError} from './invalid_pipe_argument_error';
+import {invalidPipeArgumentError} from './utils';
 import {RuntimeErrorCode} from '../errors';
 
 /**
