+++ packages/common/src/pipes/utils.ts
@@ -6,7 +6,12 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {Type, ɵRuntimeError as RuntimeError, ɵstringify as stringify} from '@angular/core';
+import {
+  Type,
+  ɵRuntimeError as RuntimeError,
+  ɵstringify as stringify,
+  isSignal,
+} from '@angular/core';
 
 import {RuntimeErrorCode} from '../errors';
 
@@ -15,4 +20,10 @@ export function invalidPipeArgumentError(type: Type<any>, value: Object) {
     RuntimeErrorCode.INVALID_PIPE_ARGUMENT,
     ngDevMode && `InvalidPipeArgument: '${value}' for pipe '${stringify(type)}'`,
   );
+}
+
+export function warnIfSignal(pipeName: string, value: unknown): void {
+  if (isSignal(value)) {
+    console.warn(`The ${pipeName} does not unwrap signals. Received a signal with value:`, value());
+  }
 }
